requirements.txt
Flask==2.1.1
Pillow==9.1.0
opencv-python==4.5.5.64